Route compose requests through the signed-in root
A mailto: tap was silently dropped on iOS whenever MessageListView
wasn't the visible view: the compose sheet was anchored to that list,
and SwiftUI cannot present a sheet from a view in a background tab.
The tick was consumed and the sheet state set, but nothing appeared —
the app just showed its last state (the App Review rejection for the
default-mail-client request) — and the orphaned sheet then popped up
whenever the user next visited the Mail tab.

Introduce ComposeRequestRouter, a ViewModifier installed once on
SignedInRootView, as the single consumer of composeRequestTick /
pendingComposeSeed. Window platforms (macOS, iPadOS, visionOS) hand
off to the compose WindowGroup as before; iPhone presents the compose
sheet from the root, which is in the visible hierarchy in every tab,
folder, and modal state. MessageListView's and MessageDetailView's
presentCompose become thin wrappers over requestCompose(seed:), so
every compose entry point funnels through the router and two sheet
hosts can never race each other.

A mailto: arriving while an iPhone compose sheet is already open no
longer replaces the draft mid-typing (the previous sheet(item:)
semantics swapped the seed out from under the user): the seed stays
parked on AppState and presents when the current sheet dismisses.

Verified in the iOS simulator against a stage account: warm open from
the Addresses tab (the reported failure), warm from the message list,
cold launch, mailto-while-composing (queued, not clobbered), and the
iPad openWindow path.
